// src/app/models/diploma.model.ts
export interface Diploma {
    diplomaId: string;
    nome: string;
    cpf: string;
    codEmecCurso: string;
    codEmecIes: string;
    codExp: string;
    codReg: string;
    dataIngresso: Date;
    dataConclusao: Date;
    dataExpDiploma: Date;
    dataRegDiploma: Date;
    dataPubDiploma: Date;
    ativo: boolean;
}